customer-support/backend/requirements.txt
Flask
flask-cors
python-dotenv
openai
requests
